+++ utils/__init__.py
@@ -3,7 +3,8 @@
 
 from .mockups import (
     crear_respuesta_negocio_no_parametrizado,
-    crear_respuesta_error_validacion
+    crear_respuesta_error_validacion,
+    crear_respuesta_preliquidacion_sin_finalizar
 )
 from .error_handlers import (
     registrar_exception_handler,
@@ -16,6 +17,7 @@ from .utils_archivos import (obtener_nombre_archivo,
 __all__ = [
     'crear_respuesta_negocio_no_parametrizado',
     'crear_respuesta_error_validacion',
+    'crear_respuesta_preliquidacion_sin_finalizar',
     'registrar_exception_handler',
     'validation_exception_handler',
     'extraer_informacion_errores',